fix: fixed section in navbaR
