+++ src/modules/core/components/NavBarCustom/sections/index.ts
@@ -1,3 +1,4 @@
 import { RolesSection } from './RolesSection'
+import { ItemLink } from './ItemLink'
 
-export { RolesSection }
+export { RolesSection, ItemLink }
